♻️ Ref : 마이페이지 리팩토링
마이페이지의 코드를 개선합니다.
